Move COH code from f3x service to report service
