+++ front-end/src/app/shared/services/report.service.ts
@@ -6,9 +6,10 @@ import { Report, ReportTypes } from '../models/report.model';
 import { TableListService } from '../interfaces/table-list-service.interface';
 import { ListRestResponse } from '../models/rest-api.model';
 import { ApiService } from './api.service';
-import { Form3X } from '../models/form-3x.model';
+import { Form3X, CashOnHand } from '../models/form-3x.model';
 import { Form24 } from '../models/form-24.model';
 import { Form99 } from '../models/form-99.model';
+import { setCashOnHandAction } from 'app/store/cash-on-hand.actions';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 function getReportFromJSON(json: any): Report {
@@ -32,6 +33,7 @@ export class ReportService implements TableListService<Report> {
     return this.apiService.get<ListRestResponse>(`${this.apiEndpoint}/?page=${pageNumber}&ordering=${ordering}`).pipe(
       map((response: ListRestResponse) => {
         response.results = response.results.map((item) => getReportFromJSON(item));
+        this.setStoreCashOnHand(response.results);
         return response;
       })
     );
@@ -89,5 +91,30 @@ export class ReportService implements TableListService<Report> {
 
   public startAmendment(report: Report): Observable<string> {
     return this.apiService.post(`${this.apiEndpoint}/${report.id}/amend/`, {});
+  }
+
+  /**
+   * Dispatches the Cash On Hand data for the first report in the list to the ngrx store.
+   * @param reports - List of reports on the current page of the Reports table
+   */
+  public setStoreCashOnHand(reports: Report[]) {
+    let payload: CashOnHand | undefined;
+
+    if (reports.length === 0) {
+      payload = {
+        report_id: undefined,
+        value: undefined,
+      };
+    } else if (reports.length > 0) {
+      const report: Form3X = reports[0] as Form3X;
+      const value = report.L6a_cash_on_hand_jan_1_ytd ?? 1.0;
+      payload = {
+        report_id: report.id,
+        value: value,
+      };
+    }
+    if (payload) {
+      this.store.dispatch(setCashOnHandAction({ payload: payload }));
+    }
   }
 }
